Добавлена интерактивная карта и слайдер выбора цены
